AURORA-139: When trying to schedule a task, only query once for active tasks in the job.
One functional change in this diff is the removal of a filter in
AttributeFilter, which was unnecessary since the query was already job-scoped.
The filter existed because at one point we supported job configurations that
could request to avoid other jobs, so the task set would include other jobs.

Bugs closed: AURORA-139

Reviewed at https://reviews.apache.org/r/17578/

Conflicts:
	src/test/java/org/apache/aurora/scheduler/state/TaskAssignerImplTest.java